+++ web/js/map/natural-events/ui.js
@@ -5,11 +5,14 @@ import markers from './markers';
 import track from './track';
 import wvui from '../../ui/ui';
 import util from '../../util/util';
-import naturalEventsUtilGetEventById from './util';
 import { CHANGE_TAB as CHANGE_SIDEBAR_TAB } from '../../modules/sidebar/constants';
 import * as EVENT_CONSTANTS from '../../modules/natural-events/constants';
 import { activateLayersForEventCategory } from '../../modules/layers/actions';
-import { deselectEvent as deselectEventAction, selected as selectedAction } from '../../modules/natural-events/actions';
+import {
+  deselectEvent as deselectEventAction,
+  selected as selectedAction,
+} from '../../modules/natural-events/actions';
+import { getDefaultEventDate, naturalEventsUtilGetEventById } from './util';
 import { selectDate } from '../../modules/date/actions';
 import { UPDATE_MAP_UI } from '../../modules/map/constants';
 import { LOCATION_POP_ACTION } from '../../redux-location-state-customs';
@@ -33,6 +36,7 @@ export default function naturalEventsUI(ui, config, store, models) {
   let naturalEventMarkers = markers(ui, store);
   const naturalEventsTracks = {};
   let naturalEventsTrack;
+
   /**
    * Suscribe to redux store and listen for
    * specific action types
@@ -56,9 +60,9 @@ export default function naturalEventsUI(ui, config, store, models) {
       case CHANGE_SIDEBAR_TAB:
         return onSidebarChange(action.activeTab);
       case EVENT_CONSTANTS.SELECT_EVENT:
-        return self.selectEvent(action.id, action.date);
+        return selectEvent(action.id, action.date);
       case EVENT_CONSTANTS.DESELECT_EVENT:
-        return self.deselectEvent();
+        return deselectEvent();
       case UPDATE_MAP_UI:
         if (map.proj !== action.ui.selected.proj) {
           onProjChange(action.ui.selected.proj);
@@ -80,7 +84,7 @@ export default function naturalEventsUI(ui, config, store, models) {
           const { selected } = events;
 
           if (selected.id) {
-            return self.selectEvent(selected.id, selected.date, null, true);
+            return selectEvent(selected.id, selected.date, null, true);
           }
         }
         break;
@@ -104,7 +108,7 @@ export default function naturalEventsUI(ui, config, store, models) {
 
       const isZoomed = Math.floor(view.getZoom()) >= 3;
       if (isZoomed || proj.selected.id !== 'geographic') {
-        self.filterEventList();
+        filterEventList();
       }
       // check if selected event is in changed projection
       if (self.selected.id) {
@@ -124,8 +128,8 @@ export default function naturalEventsUI(ui, config, store, models) {
         });
         // remove selected event if not in changed projection
         if (!findSelectedInProjection) {
-          self.deselectEvent();
-          self.filterEventList();
+          deselectEvent();
+          filterEventList();
         }
       }
     } else {
@@ -147,11 +151,11 @@ export default function naturalEventsUI(ui, config, store, models) {
     if (naturalEventsTrack.trackDetails.id) naturalEventsTrack.update(null);
     naturalEventsTrack = naturalEventsTracks[id];
     // filter events within projection view extent
-    self.filterEventList();
+    filterEventList();
 
     // handle list filter on map move
     map.on('moveend', (e) => {
-      self.filterEventList();
+      filterEventList();
     });
 
     if (state.sidebar.activeTab === 'events') {
@@ -159,7 +163,7 @@ export default function naturalEventsUI(ui, config, store, models) {
       naturalEventMarkers.remove(self.markers);
       // Store markers so the can be referenced later
       self.markers = naturalEventMarkers.draw();
-      self.filterEventList();
+      filterEventList();
       // check if selected event is in changed projection
       if (self.selected.id) {
         const findSelectedInProjection = lodashFind(self.markers, (
@@ -178,16 +182,17 @@ export default function naturalEventsUI(ui, config, store, models) {
         });
         // remove selected event if not in changed projection
         if (!findSelectedInProjection) {
-          self.deselectEvent();
-          self.filterEventList();
+          deselectEvent();
+          filterEventList();
         }
         if (self.selected.date) {
           const event = naturalEventsUtilGetEventById(self.eventsData, self.selected.id);
-          naturalEventsTrack.update(event, self.selected.date, self.selectEvent);
+          naturalEventsTrack.update(event, self.selected.date, selectEvent);
         }
       }
     }
   };
+
   const onQueryResults = function() {
     const state = store.getState();
 
@@ -200,19 +205,19 @@ export default function naturalEventsUI(ui, config, store, models) {
 
     const isZoomed = Math.floor(view.getZoom()) >= 3;
     if (isZoomed || state.proj.selected.id !== 'geographic') {
-      self.filterEventList();
+      filterEventList();
     }
 
     map.on('moveend', (e) => {
       const isZoomed = Math.floor(view.getZoom()) >= 3;
       if (isZoomed || state.proj.selected.id !== 'geographic') {
-        self.filterEventList();
+        filterEventList();
       }
     });
 
     // Reselect previously selected event
     if (self.selected.id) {
-      self.selectEvent(
+      selectEvent(
         self.selected.id,
         self.selected.date || null,
         null,
@@ -220,6 +225,7 @@ export default function naturalEventsUI(ui, config, store, models) {
       );
     }
   };
+
   const init = function() {
     map = ui.map.selected;
     lodashForOwn(ui.map.proj, (projMap, key) => {
@@ -251,7 +257,7 @@ export default function naturalEventsUI(ui, config, store, models) {
    * @param {*} rotation
    * @param {*} isInitialLoad
    */
-  self.selectEvent = function(id, date, rotation, isInitialLoad) {
+  const selectEvent = function(id, date, rotation, isInitialLoad) {
     const isIdChange = !self.selected || self.selected.id !== id;
     const prevId = self.selected.id ? self.selected.id : false;
     const prevEvent = prevId && naturalEventsUtilGetEventById(self.eventsData, prevId);
@@ -263,7 +269,7 @@ export default function naturalEventsUI(ui, config, store, models) {
       wvui.notify(`The event with an id of ${id} is no longer active.`);
       return;
     }
-    date = date || self.getDefaultEventDate(event);
+    date = date || getDefaultEventDate(event);
     self.selected = {
       id,
       date,
@@ -311,12 +317,12 @@ export default function naturalEventsUI(ui, config, store, models) {
       } else {
         ui.map.updateDate();
       }
-      naturalEventsTrack.update(event, date, self.selectEvent);
+      naturalEventsTrack.update(event, date, selectEvent);
       store.dispatch(selectedAction());
     });
   };
 
-  self.deselectEvent = function() {
+  const deselectEvent = function() {
     self.selected = {};
     naturalEventMarkers.remove(self.markers);
     const state = store.getState();
@@ -328,24 +334,12 @@ export default function naturalEventsUI(ui, config, store, models) {
     self.events.trigger('change');
   };
 
-  self.getDefaultEventDate = function(event) {
-    let date = new Date(event.geometries[0].date).toISOString().split('T')[0];
-    if (event.geometries.length < 2) return date;
-    const category = event.categories.title || event.categories[0].title;
-    const today = new Date().toISOString().split('T')[0];
-    // For storms that happened today, get previous date
-    if (date === today && category === 'Severe Storms') {
-      [date] = new Date(event.geometries[1].date).toISOString().split('T');
-    }
-    return date;
-  };
-
   /**
    * Filter event list in sidebar based on projection/view extents
    *
    * @param  {Boolean} showAll - show all available points in projection
    */
-  self.filterEventList = function() {
+  const filterEventList = function() {
     const state = store.getState();
     const { proj } = state;
     const { showAll } = state.events;
@@ -358,11 +352,11 @@ export default function naturalEventsUI(ui, config, store, models) {
 
     self.eventsData.forEach((naturalEvent) => {
       const isSelectedEvent = self.selected.id === naturalEvent.id;
-      let date = self.getDefaultEventDate(naturalEvent);
+      let date = getDefaultEventDate(naturalEvent);
       if (self.selected && self.selected.date) {
         date = self.selected.date;
       }
-      const geometry = lodashFind(naturalEvent.geometries, (geometry) => geometry.date.split('T')[0] === date) || naturalEvent.geometries[0];
+      const geometry = lodashFind(naturalEvent.geometry, (geometry) => geometry.date.split('T')[0] === date) || naturalEvent.geometry[0];
 
       let { coordinates } = geometry;
 
@@ -425,7 +419,7 @@ export default function naturalEventsUI(ui, config, store, models) {
     const zoom = isSameEventID
       ? ui.map.selected.getView().getZoom()
       : zoomLevelReference[category];
-    const geometry = lodashFind(event.geometries, (geom) => geom.date.split('T')[0] === date);
+    const geometry = lodashFind(event.geometry, (geom) => geom.date.split('T')[0] === date);
 
     // check for polygon geometries and/or perform projection coordinate transform
     let coordinates = geometry.type === 'Polygon'
